Solutions/AzureSubscriptionOwnerAudit/requirements.txt
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==20.0.0
azure-mgmt-subscription==3.1.1
tabulate==0.9.0
azure_graphrbac==0.61.2